# apps/merchant-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/inventory": {
      "get": {
        "operationId": "InventoryController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "SKU or product name match",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "locationId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "stockLte",
            "required": false,
            "in": "query",
            "description": "only rows at or below this on-hand quantity",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedInventory"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Inventory"
        ]
      }
    },
    "/inventory/movements": {
      "post": {
        "operationId": "InventoryController_createMovement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryMovementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryMovementRecord"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Inventory"
        ]
      },
      "get": {
        "operationId": "InventoryController_findMovements",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "variantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "locationId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "reason",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "received",
                "sold",
                "return",
                "damaged",
                "adjustment"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedInventoryMovements"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Inventory"
        ]
      }
    },
    "/locations": {
      "post": {
        "operationId": "LocationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Locations"
        ]
      },
      "get": {
        "operationId": "LocationsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "name match",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLocations"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Locations"
        ]
      }
    },
    "/locations/{id}": {
      "patch": {
        "operationId": "LocationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Locations"
        ]
      },
      "delete": {
        "operationId": "LocationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Locations"
        ]
      }
    },
...
    "/carts": {
      "get": {
        "operationId": "CartsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCarts"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Carts"
        ]
      }
    },
    "/carts/{id}": {
      "get": {
        "operationId": "CartsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Carts"
        ]
      }
    },
...
    "/pos-devices": {
      "post": {
        "operationId": "PosDevicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePosDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosDevicePairing"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PosDevices"
        ]
      },
      "get": {
        "operationId": "PosDevicesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosDevice"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PosDevices"
        ]
      }
    },
    "/pos-devices/{id}": {
      "patch": {
        "operationId": "PosDevicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePosDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosDevice"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PosDevices"
        ]
      }
    },
    "/pos-devices/{id}/revoke": {
      "post": {
        "operationId": "PosDevicesController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosDevice"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PosDevices"
        ]
      }
    },
    "/pos-devices/{id}/rotate-pairing": {
      "post": {
        "operationId": "PosDevicesController_rotatePairing",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosDevicePairing"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "PosDevices"
        ]
      }
    }
  },
  "info": {
    "title": "Merchant",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InventoryRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "variantId": {
            "type": "number"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "productId": {
            "type": "number"
          },
          "productName": {
            "type": "string"
          },
          "locationId": {
            "type": "number"
          },
          "locationName": {
            "type": "string"
          },
          "stock": {
            "type": "number"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "variantId",
          "sku",
          "productId",
          "productName",
          "locationId",
          "locationName",
          "stock",
          "updatedAt"
        ]
      },
      "PaginatedInventory": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryRecord"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "CreateInventoryMovementDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "number"
          },
          "locationId": {
            "type": "number"
          },
          "delta": {
            "type": "number"
          },
          "reason": {
            "type": "string",
            "enum": [
              "received",
              "sold",
              "return",
              "damaged",
              "adjustment"
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "variantId",
          "locationId",
          "delta",
          "reason"
        ]
      },
      "InventoryMovementRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "variantId": {
            "type": "number"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "productId": {
            "type": "number"
          },
          "productName": {
            "type": "string"
          },
          "locationId": {
            "type": "number"
          },
          "locationName": {
            "type": "string"
          },
          "delta": {
            "type": "number"
          },
          "reason": {
            "type": "string",
            "enum": [
              "received",
              "sold",
              "return",
              "damaged",
              "adjustment"
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdByEmail": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "variantId",
          "sku",
          "productId",
          "productName",
          "locationId",
          "locationName",
          "delta",
          "reason",
          "note",
          "createdByEmail",
          "createdAt"
        ]
      },
      "PaginatedInventoryMovements": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryMovementRecord"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "CreateLocationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "accountId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "addressCity": {
            "type": "string",
            "nullable": true
          },
          "addressState": {
            "type": "string",
            "nullable": true
          },
          "addressPostalCode": {
            "type": "string",
            "nullable": true
          },
          "addressCountry": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "accountId",
          "name",
          "addressLine1",
          "addressLine2",
          "addressCity",
          "addressState",
          "addressPostalCode",
          "addressCountry",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedLocations": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Location"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "addressCity": {
            "type": "string",
            "nullable": true
          },
          "addressState": {
            "type": "string",
            "nullable": true
          },
          "addressPostalCode": {
            "type": "string",
            "nullable": true
          },
          "addressCountry": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CartListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "token": {
            "type": "string"
          },
          "itemCount": {
            "type": "number"
          },
          "subtotalCents": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "token",
          "itemCount",
          "subtotalCents",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedCarts": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartListItem"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "CartDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "token": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartDetailItem"
            }
          },
          "subtotalCents": {
            "type": "number"
          },
          "itemCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "token",
          "createdAt",
          "updatedAt",
          "items",
          "subtotalCents",
          "itemCount"
        ]
      },
...
      "CreatePosDeviceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "locationId": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "locationId"
        ]
      },
      "PosDevicePairing": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "locationId": {
            "type": "number"
          },
          "pairingCode": {
            "type": "string"
          },
          "pairingExpiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "locationId",
          "pairingCode",
          "pairingExpiresAt"
        ]
      },
      "PosDevice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "locationId": {
            "type": "number"
          },
          "locationName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "revoked"
            ]
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "pairedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "revokedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "locationId",
          "locationName",
          "status",
          "lastSeenAt",
          "pairedAt",
          "revokedAt",
          "createdAt"
        ]
      },
      "UpdatePosDeviceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "locationId": {
            "type": "number"
          }
        }
      }
    }
  }
}